Remove now-unneeded mod type annotations
Most of the mod type annotations can now be changed to MOD_BOTH, because
if you pass a mask which can only contain real mods in the first place to
e.g. ModMaskText, then MOD_REAL and MOD_BOTH will give the same result.

In the cases where MOD_BOTH is only ever the argument, we just remove
it. What's left is where it really "matters".

Signed-off-by: Ran Benita <ran234@gmail.com>