compat: fix bad interpret predicate mods "all" calculation
Commit 9984d1d03cd78eb636c75cc2bbd2d240dc1dd72f changed the type of
interpret->mods to xkb_mod_mask_t, but this bit of code assumes that the
type is uint8_t.
This code is not usually run (for example by our tests), but when it
does keymap-dump would print out all of the modifiers (including the
virtual ones) which causes recompilation of the output to fail
miserably.

https://bugs.freedesktop.org/show_bug.cgi?id=55769

Signed-off-by: Ran Benita <ran234@gmail.com>