keycodes: remove unneeded alias conflict check
This is already checked when adding a new alias and merging aliases, so
it can never happen when we get to copying to the keymap.

Also the log verbosity decision there is quite useless, we should just
warn always and be done with it. So we can remove the file_id from
AliasInfo, and collapse the alias functions together.

Signed-off-by: Ran Benita <ran234@gmail.com>