parser: don't use %locations
byacc doesn't support this feature.

We print the line/col of the last scanned token instead. This is slightly
less in case of *parser* errors (not syntax errors), but I couldn't make
it point to another line, and this are pretty cryptic anyways. So it's
good enough. Also might be a bit faster, but haven't checked.

Signed-off-by: Ran Benita <ran234@gmail.com>